companion-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
qdrant-client==1.7.1
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9